Fix CID 1475227 1475229 1475231 1475237 1475245 1475246 1475251 1475257 1475278 1475282

+++ sqrt/crtalgsqrt.c
@@ -1019,7 +1019,7 @@ struct sqrt_globals {
     int s;      // number of shares of A
     int t;      // number of prime groups
     int r;      // number of primes in each prime group
-    int prec;
+    unsigned long prec;
     mpz_t P;    // prime product (not to the power prec)
     size_t nbits_sqrt;
     // size_t nbits_a;
@@ -2365,7 +2365,7 @@ void * precompute_powers_child(struct subtask_info_t * info)/* {{{ */
 {
     struct prime_data * p = info->p;
 
-    logprint("Precomputing p^%d, p=%lu\n", glob.prec, p->p);
+    logprint("Precomputing p^%lu, p=%lu\n", glob.prec, p->p);
     // this triggers the whole precomputation.
     power_lookup(p->powers, glob.prec);
 
@@ -3588,7 +3588,7 @@ int main(int argc, char **argv)
 
     if (glob.rank == 0) {
         char sbuf[32];
-        fprintf(stderr, "# [%2.2lf] Lifting to precision l=%d (p^l is approx %s)\n", WCT, glob.prec, size_disp(glob.prec * log(primes[0].p)/M_LN2 / 8, sbuf));
+        fprintf(stderr, "# [%2.2lf] Lifting to precision l=%lu (p^l is approx %s)\n", WCT, glob.prec, size_disp(glob.prec * log(primes[0].p)/M_LN2 / 8, sbuf));
     }
 
     mpi_set_communicators();
